+++ src/main/java/net/mcreator/jujutsucraftaddon/client/particle/CursedPowerOkkotsuParticle.java
@@ -1,66 +1,60 @@
-
 package net.mcreator.jujutsucraftaddon.client.particle;
 
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.api.distmarker.Dist;
-
-import net.minecraft.core.particles.SimpleParticleType;
-import net.minecraft.client.particle.TextureSheetParticle;
-import net.minecraft.client.particle.SpriteSet;
-import net.minecraft.client.particle.ParticleRenderType;
-import net.minecraft.client.particle.ParticleProvider;
-import net.minecraft.client.particle.Particle;
 import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.particle.*;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 
 @OnlyIn(Dist.CLIENT)
 public class CursedPowerOkkotsuParticle extends TextureSheetParticle {
-	public static CursedPowerOkkotsuParticleProvider provider(SpriteSet spriteSet) {
-		return new CursedPowerOkkotsuParticleProvider(spriteSet);
-	}
+    public static CursedPowerOkkotsuParticleProvider provider(SpriteSet spriteSet) {
+        return new CursedPowerOkkotsuParticleProvider(spriteSet);
+    }
 
-	public static class CursedPowerOkkotsuParticleProvider implements ParticleProvider<SimpleParticleType> {
-		private final SpriteSet spriteSet;
+    public static class CursedPowerOkkotsuParticleProvider implements ParticleProvider<SimpleParticleType> {
+        private final SpriteSet spriteSet;
 
-		public CursedPowerOkkotsuParticleProvider(SpriteSet spriteSet) {
-			this.spriteSet = spriteSet;
-		}
+        public CursedPowerOkkotsuParticleProvider(SpriteSet spriteSet) {
+            this.spriteSet = spriteSet;
+        }
 
-		public Particle createParticle(SimpleParticleType typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
-			return new CursedPowerOkkotsuParticle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed, this.spriteSet);
-		}
-	}
+        public Particle createParticle(SimpleParticleType typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
+            return new CursedPowerOkkotsuParticle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed, this.spriteSet);
+        }
+    }
 
-	private final SpriteSet spriteSet;
+    private final SpriteSet spriteSet;
 
-	protected CursedPowerOkkotsuParticle(ClientLevel world, double x, double y, double z, double vx, double vy, double vz, SpriteSet spriteSet) {
-		super(world, x, y, z);
-		this.spriteSet = spriteSet;
-		this.setSize(0.5f, 0.5f);
-		this.quadSize *= 3f;
-		this.lifetime = (int) Math.max(1, 20 + (this.random.nextInt(4) - 2));
-		this.gravity = 0f;
-		this.hasPhysics = true;
-		this.xd = vx * 0;
-		this.yd = vy * 0;
-		this.zd = vz * 0;
-		this.setSpriteFromAge(spriteSet);
-	}
+    protected CursedPowerOkkotsuParticle(ClientLevel world, double x, double y, double z, double vx, double vy, double vz, SpriteSet spriteSet) {
+        super(world, x, y, z);
+        this.spriteSet = spriteSet;
+        this.setSize(0.5f, 0.5f);
+        this.quadSize *= 3f;
+        this.lifetime = Math.max(1, 20 + (this.random.nextInt(4) - 2));
+        this.gravity = 0f;
+        this.hasPhysics = true;
+        this.xd = vx * 0;
+        this.yd = vy * 0;
+        this.zd = vz * 0;
+        this.setSpriteFromAge(spriteSet);
+    }
 
-	@Override
-	public int getLightColor(float partialTick) {
-		return 15728880;
-	}
+    @Override
+    public int getLightColor(float partialTick) {
+        return 15728880;
+    }
 
-	@Override
-	public ParticleRenderType getRenderType() {
-		return ParticleRenderType.PARTICLE_SHEET_LIT;
-	}
+    @Override
+    public ParticleRenderType getRenderType() {
+        return ParticleRenderType.PARTICLE_SHEET_LIT;
+    }
 
-	@Override
-	public void tick() {
-		super.tick();
-		if (!this.removed) {
-			this.setSprite(this.spriteSet.get((this.age / 2) % 10 + 1, 10));
-		}
-	}
+    @Override
+    public void tick() {
+        super.tick();
+        if (!this.removed) {
+            this.setSprite(this.spriteSet.get((this.age / 2) % 10 + 1, 10));
+        }
+    }
 }
