+++ src/main/java/net/mcreator/jujutsucraftaddon/client/particle/HorizontalSlash1Particle.java
@@ -1,66 +1,60 @@
-
 package net.mcreator.jujutsucraftaddon.client.particle;
 
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.api.distmarker.Dist;
-
-import net.minecraft.core.particles.SimpleParticleType;
-import net.minecraft.client.particle.TextureSheetParticle;
-import net.minecraft.client.particle.SpriteSet;
-import net.minecraft.client.particle.ParticleRenderType;
-import net.minecraft.client.particle.ParticleProvider;
-import net.minecraft.client.particle.Particle;
 import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.particle.*;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 
 @OnlyIn(Dist.CLIENT)
 public class HorizontalSlash1Particle extends TextureSheetParticle {
-	public static HorizontalSlash1ParticleProvider provider(SpriteSet spriteSet) {
-		return new HorizontalSlash1ParticleProvider(spriteSet);
-	}
+    public static HorizontalSlash1ParticleProvider provider(SpriteSet spriteSet) {
+        return new HorizontalSlash1ParticleProvider(spriteSet);
+    }
 
-	public static class HorizontalSlash1ParticleProvider implements ParticleProvider<SimpleParticleType> {
-		private final SpriteSet spriteSet;
+    public static class HorizontalSlash1ParticleProvider implements ParticleProvider<SimpleParticleType> {
+        private final SpriteSet spriteSet;
 
-		public HorizontalSlash1ParticleProvider(SpriteSet spriteSet) {
-			this.spriteSet = spriteSet;
-		}
+        public HorizontalSlash1ParticleProvider(SpriteSet spriteSet) {
+            this.spriteSet = spriteSet;
+        }
 
-		public Particle createParticle(SimpleParticleType typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
-			return new HorizontalSlash1Particle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed, this.spriteSet);
-		}
-	}
+        public Particle createParticle(SimpleParticleType typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
+            return new HorizontalSlash1Particle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed, this.spriteSet);
+        }
+    }
 
-	private final SpriteSet spriteSet;
+    private final SpriteSet spriteSet;
 
-	protected HorizontalSlash1Particle(ClientLevel world, double x, double y, double z, double vx, double vy, double vz, SpriteSet spriteSet) {
-		super(world, x, y, z);
-		this.spriteSet = spriteSet;
-		this.setSize(0.2f, 0.2f);
-		this.quadSize *= 10f;
-		this.lifetime = (int) Math.max(1, 20 + (this.random.nextInt(6) - 3));
-		this.gravity = 0f;
-		this.hasPhysics = true;
-		this.xd = vx * 0;
-		this.yd = vy * 0;
-		this.zd = vz * 0;
-		this.setSpriteFromAge(spriteSet);
-	}
+    protected HorizontalSlash1Particle(ClientLevel world, double x, double y, double z, double vx, double vy, double vz, SpriteSet spriteSet) {
+        super(world, x, y, z);
+        this.spriteSet = spriteSet;
+        this.setSize(0.2f, 0.2f);
+        this.quadSize *= 10f;
+        this.lifetime = Math.max(1, 20 + (this.random.nextInt(6) - 3));
+        this.gravity = 0f;
+        this.hasPhysics = true;
+        this.xd = vx * 0;
+        this.yd = vy * 0;
+        this.zd = vz * 0;
+        this.setSpriteFromAge(spriteSet);
+    }
 
-	@Override
-	public int getLightColor(float partialTick) {
-		return 15728880;
-	}
+    @Override
+    public int getLightColor(float partialTick) {
+        return 15728880;
+    }
 
-	@Override
-	public ParticleRenderType getRenderType() {
-		return ParticleRenderType.PARTICLE_SHEET_LIT;
-	}
+    @Override
+    public ParticleRenderType getRenderType() {
+        return ParticleRenderType.PARTICLE_SHEET_LIT;
+    }
 
-	@Override
-	public void tick() {
-		super.tick();
-		if (!this.removed) {
-			this.setSprite(this.spriteSet.get((this.age / 2) % 8 + 1, 8));
-		}
-	}
+    @Override
+    public void tick() {
+        super.tick();
+        if (!this.removed) {
+            this.setSprite(this.spriteSet.get((this.age / 2) % 8 + 1, 8));
+        }
+    }
 }
