+++ src/main/java/net/mcreator/jujutsucraftaddon/procedures/HealthScrollRightclickedProcedure.java
@@ -1,35 +1,34 @@
 package net.mcreator.jujutsucraftaddon.procedures;
 
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.network.chat.Component;
-
-import net.mcreator.jujutsucraftaddon.network.JujutsucraftaddonModVariables;
 import net.mcreator.jujutsucraftaddon.init.JujutsucraftaddonModItems;
+import net.mcreator.jujutsucraftaddon.network.JujutsucraftaddonModVariables;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 
 public class HealthScrollRightclickedProcedure {
-	public static void execute(Entity entity) {
-		if (entity == null)
-			return;
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(
-					new MobEffectInstance(MobEffects.HEALTH_BOOST, -1, (int) ((entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.HEALTH_BOOST) ? _livEnt.getEffect(MobEffects.HEALTH_BOOST).getAmplifier() : 0) + 1), false, false));
-		{
-			double _setval = (entity.getCapability(JujutsucraftaddonModVariables.PLAYER_VARIABLES_CAPABILITY, null).orElse(new JujutsucraftaddonModVariables.PlayerVariables())).HealthAttribute;
-			entity.getCapability(JujutsucraftaddonModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
-				capability.HealthAttribute = _setval;
-				capability.syncPlayerVariables(entity);
-			});
-		}
-		if (entity instanceof Player _player && !_player.level().isClientSide())
-			_player.displayClientMessage(Component.literal("Added Health Points"), false);
-		if (entity instanceof Player _player) {
-			ItemStack _stktoremove = new ItemStack(JujutsucraftaddonModItems.HEALTH_SCROLL.get());
-			_player.getInventory().clearOrCountMatchingItems(p -> _stktoremove.getItem() == p.getItem(), 1, _player.inventoryMenu.getCraftSlots());
-		}
-	}
+    public static void execute(Entity entity) {
+        if (entity == null)
+            return;
+        if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(
+                    new MobEffectInstance(MobEffects.HEALTH_BOOST, -1, (entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.HEALTH_BOOST) ? _livEnt.getEffect(MobEffects.HEALTH_BOOST).getAmplifier() : 0) + 1, false, false));
+        {
+            double _setval = (entity.getCapability(JujutsucraftaddonModVariables.PLAYER_VARIABLES_CAPABILITY, null).orElse(new JujutsucraftaddonModVariables.PlayerVariables())).HealthAttribute;
+            entity.getCapability(JujutsucraftaddonModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
+                capability.HealthAttribute = _setval;
+                capability.syncPlayerVariables(entity);
+            });
+        }
+        if (entity instanceof Player _player && !_player.level().isClientSide())
+            _player.displayClientMessage(Component.literal("Added Health Points"), false);
+        if (entity instanceof Player _player) {
+            ItemStack _stktoremove = new ItemStack(JujutsucraftaddonModItems.HEALTH_SCROLL.get());
+            _player.getInventory().clearOrCountMatchingItems(p -> _stktoremove.getItem() == p.getItem(), 1, _player.inventoryMenu.getCraftSlots());
+        }
+    }
 }
