+++ src/main/java/net/mcreator/jujutsucraftaddon/procedures/MakiiNbtProcedure.java
@@ -1,36 +1,35 @@
 package net.mcreator.jujutsucraftaddon.procedures;
 
-import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.commands.CommandSourceStack;
-import net.minecraft.commands.CommandSource;
-
 import net.mcreator.jujutsucraft.init.JujutsucraftModMobEffects;
+import net.minecraft.commands.CommandSource;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.level.LevelAccessor;
 
 public class MakiiNbtProcedure {
-	public static void execute(LevelAccessor world, Entity entity) {
-		if (entity == null)
-			return;
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(MobEffects.DAMAGE_BOOST, -1, 25, false, false));
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, -1, 5, false, false));
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, -1, 3, false, false));
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get(), (-1),
-					(int) ((entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get()) ? _livEnt.getEffect(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get()).getAmplifier() : 0) + 3),
-					false, false));
-		{
-			Entity _ent = entity;
-			if (!_ent.level().isClientSide() && _ent.getServer() != null) {
-				_ent.getServer().getCommands().performPrefixedCommand(new CommandSourceStack(CommandSource.NULL, _ent.position(), _ent.getRotationVector(), _ent.level() instanceof ServerLevel ? (ServerLevel) _ent.level() : null, 4,
-						_ent.getName().getString(), _ent.getDisplayName(), _ent.level().getServer(), _ent), "execute as @s run item replace entity @s weapon.mainhand with jujutsucraft:dragon_bone");
-			}
-		}
-	}
+    public static void execute(LevelAccessor world, Entity entity) {
+        if (entity == null)
+            return;
+        if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(new MobEffectInstance(MobEffects.DAMAGE_BOOST, -1, 25, false, false));
+        if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, -1, 5, false, false));
+        if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, -1, 3, false, false));
+        if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(new MobEffectInstance(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get(), (-1),
+                    (entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get()) ? _livEnt.getEffect(JujutsucraftModMobEffects.PHYSICAL_GIFTED_EFFECT.get()).getAmplifier() : 0) + 3,
+                    false, false));
+        {
+            Entity _ent = entity;
+            if (!_ent.level().isClientSide() && _ent.getServer() != null) {
+                _ent.getServer().getCommands().performPrefixedCommand(new CommandSourceStack(CommandSource.NULL, _ent.position(), _ent.getRotationVector(), _ent.level() instanceof ServerLevel ? (ServerLevel) _ent.level() : null, 4,
+                        _ent.getName().getString(), _ent.getDisplayName(), _ent.level().getServer(), _ent), "execute as @s run item replace entity @s weapon.mainhand with jujutsucraft:dragon_bone");
+            }
+        }
+    }
 }
