+++ src/main/java/net/mcreator/jujutsucraftaddon/procedures/NaoyaBuffsProcedure.java
@@ -1,16 +1,16 @@
 package net.mcreator.jujutsucraftaddon.procedures;
 
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 
 public class NaoyaBuffsProcedure {
-	public static void execute(Entity sourceentity) {
-		if (sourceentity == null)
-			return;
-		if (sourceentity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 120,
-					(int) ((sourceentity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.MOVEMENT_SPEED) ? _livEnt.getEffect(MobEffects.MOVEMENT_SPEED).getAmplifier() : 0) + 1), false, false));
-	}
+    public static void execute(Entity sourceentity) {
+        if (sourceentity == null)
+            return;
+        if (sourceentity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+            _entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 120,
+                    (sourceentity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.MOVEMENT_SPEED) ? _livEnt.getEffect(MobEffects.MOVEMENT_SPEED).getAmplifier() : 0) + 1, false, false));
+    }
 }
